refactor(schedule-repeat): common 구조 수정 후 테스트코드 수정
수정이 되지 않은 테스트 코드 수정
